example+wasm-client: close connection on page unload
To make sure the connection is closed when the page is unloaded (browser
or tab is closed or the user navigates away from Terminal Web), we add
an onunload hook to the script.
This will make it more reliable to connect again since the server side
can properly close the connection on its end as well.