mailbox: implement Hash2Curve
We need generator points M and N for blinding our server and client
public keys. We choose points on the secp256k1 curve that nobody knows
the discrete log of by hashing constant strings to the curve, using the
reference implementation in https://github.com/armfazh/h2c-go-ref.
The implementation mentions that the code should not be used in
production, as it's a reference implementation only. But for our
purposes of just blinding another public key this should be sufficiently
safe to use.