mailbox/numsgen: add new CLI utility to generate NUMs points for SPAKE2
In this commit, we add a new utility the generates NUMs points using a
try-and-increment approach. The starting phrase we use by default to
generate a point is "Lightning Connect". This will be used in later
commits to generate a NUMs point for our Noise variant that uses masked
ephemeral points with SPAKE2 as the public key masking operation.