mailbox: propagate rename to NUMS point and Noise prologue
In this commit, we propagete the rename of the project to the initial
string used to generate the NUMs point as well as the string used for
the Noise prologue. We also regnerate the NUMs point using the new
default values and also update the hard coded value in the noise repo.