mailbox: handshake versioning logic and test
In this commit, the idea of a min supported and max supported handshake
version is introduced so that client and server can negotiate which
version they will use. The client will always send the minimum version
they support in act 1. The server will then send its maximum supported
version back in act 2. This is done due to the fact that for TW, the
client will always be the first to upgrade versions. What ever the
server sends back in act 2 is the version they will continue with.

This commit also adds a unit test to test if a handshake is successfull
given different version congigurations.