mailbox: add handshake version 1
In this commit, HanshakeVersion1 is added. If this version is decided
on, then the auth data is sent from server to client in a normal message
after act 3 is complete.

The auth data can exceed the packet size limit of 2^16 and so two new
functions are added to `Machine`: `WriteAuthData` which prepends the
authdata with its length and then sends the payload in as many chunks as
necessary. `ReadAuthData` reads the expected payload length and then
reads the appropriate number of chunks.