mailbox: use receive buffer to allow multiple calls
GBN reads an entire packet at once and so does not really care about the
number of bytes that the caller is expecting to read. To allow the
caller to read a payload byte by byte, we can use a bytes buffer which
we will populate with the full gbn payload and then subsequent read
calls will read from the buffer until it is empty.