gbn: NACK handling for successful queue send
In the case where we send the entire queue successfully but just did not
get any ACKs in return, when we resend the queue, the other side will
send back NACK with a sequence number that is _not_ in our queue but is
the next sequence number we will send. This means that if we get a NACK
with a sequence number equal to our seqTop, then we can empty the queue
and assume that it was all successfully sent.