gbn: add recv and send timeouts
In this commit, we add send and receive timout tickers to the GBN Send
and Recv functions as well as methods that can be used to set the
timeouts used for these. A unit test is also added to ensure that the
clients 'Recv' function does correcly error out on timeout.