mailbox: ensure proper client refresh
Currently, grpc is passed the ClientConn itself meaning that after a
retry, the closeOnce sync var is never called again resulting in an
incorrect refresh. This commit wraps the ClientConn in a Client object
(similarly to how the ServerConn is wrapped by a Server) so that the
Client can properly refresh the ClientConn each time the client
re-Dials.