itest+wasm-client: add WithBlock dial option to client
In this commit, we add a WithBlock client dial option so that the
connection is not returned until the connection has been established
(ie, handshake is complete). The wasm-client is also updated to
spin the `mailboxRPCConnection` off into a goroutine. This is
required due to the function being called from within a `js.FuncOf`
function (https://pkg.go.dev/syscall/js#FuncOf).