mod: Bump lightning-terminal, lnd & grpc versions
Bump lightning-terminal to version v0.10.1-alpha. Since we also
specifically like to include a fix that was merged in the master branch
of lightning-terminal after the v0.10.1-alpha release, we bump the
version to the commit of the master branch that includes that fix
(v0.10.1-alpha.0.20230612093826-ee04747e4a34). This fix is required in
order to be able to add taproot-assets support for the WASM client.

Also update lnd to version v0.16.3-beta and grpc to version v1.45.0, as
it's required with the lightning-terminal bump.