mod: bump wasm-client LiT to `v0.10.5-alpha`
Bump the lightning-terminal dependency for the wasm-client to
`v0.10.5-alpha`. Also bump lnd to `v0.16.4-beta`, taproot-assets to
`v0.2.3`, loop to `v0.26.2-beta` and grpc to `v1.53.0` as it's required
with the lightning-terminal bump. Also bump the lightning-node-connect
dependency to `v0.2.4-alpha`, but as this dependency is replaced, this
doesn't have any real effect.