gbn: add timeout manager
The timeout manager is a manager for all timeout values used by the
different components of the gbn package. The timeout manager enables
the functionality of dynamically changing the resend timeout value based
on how long it takes to receive an response from the counterparty.

To make the gbn package more modular, we'll move the responsibility of
managing the timeout values of the gbn package to the timeout manager in
the next commit.