gbn: change implication of processNACK signature
The processNACK function returns two booleans as its signature.
The first boolean used to only indicate if the NACK was part of the
queue, but is with this commit expanded to indicate if the queue needs
to be resent or not as a result of the NACK.

The second boolean indicates if the queue sequence base was bumped or
not. Previously, the bumped value for processNACK didn't accurately
reflect if the queue sequence base was actually bumped or not for the
cases when we didn't trigger a resend. This commit addresses that issue.
Updating the bumped value to be correct has no real impact on the
current implementation of the gbn conn, but it might start having an
effect if the implementation is modified in the future.