gbn: use timeout manager to handle timeouts
This commit moves the responsibility of managing the timeout values
throughout the gbn package to the `TimeoutManager` struct.

Even though the `TimeoutManager` is now used to manage timeouts, the
resend timeout is not yet set dynamically for the connections of the gbn
package. Support and usage of that functionality will added in the
upcoming commits.