gbn: support dynamic resend timeouts
This commit adds support for setting the resend timeout dynamically for
a connection. Prior to this commit, the timeout before a client or
server resends the queue of packets, was always set to fixed value.
A fixed timeout isn't suitable for all connections though, as the
latency for different clients varies.

With this commit, we instead add support for setting the resend timeout
based on:
* How long it took for the other party to respond during the
handshake process
* How long it took for the other party to respond with the connect ACK
for a sent data packet.

The timeout is set then to the time it took for the server to respond,
multiplied by the resendMultiplier, unless the duration is shorter than
the default resend timeout.

Note though that if a connection's resend timeout is set manually set
through the WithStaticResendTimeout, the resend timeout will always be
set to that value, and won't be dynamically updated.