gbn: complete handshake if client has completed it
This fixes a bug where the server would not complete the handshake if
the server restarted the handshake after sending it's SYN, which the
client then received and then completed the handshake.
This could previously happen if the server timedout when waiting for the
clients SYN response, or if the client's SYNACK was lost due to packet
loss.