wasm-client: close client conn in a goroutine
Close the client connection in a goroutine. This fixes a bug that was
introduced with the upgrade to grpc v0.18.1. The caused the websocket
to freeze during closure when the client disconnected, and therefore
the entire wasm-client would freeze.
When closing the connection in a goroutine, the websocket is closed
correctly.