ci: add GitHub action(s) to clean up runner disk space
Define composite actions to remove large toolsets, unused packages,
Docker images, and caches from the GitHub runner, freeing up disk space
for builds.

Note that the content of the cleanup-space/actions.yaml file is copied
from the following `tapd` PR:
https://github.com/lightninglabs/taproot-assets/pull/1878

There's one distinct difference from the `tapd` file. In this repo,
we separate the cleanup of the android folder into a separate action.
The reason for this separation is that we need to be able to control
when the android folder is cleaned up, during the job, as some jobs
will fail if the android folder is removed in the beginning of the job.

Not all of the targeted files when running the script, should ever
have been used when building the binaries for LNC, but we keep the
script consistent between the repos for simplicity.