ci: execute cleanup commands in CI jobs
Update workflows to use the new `cleanup-space` composite action for
clearing runner disk space.

Note that in the "build package and wasm" job, the cleanup of the
android build folder is done after the rest of the step is completed, as
the contents of the folder are needed during the "android compile" step.